building out profile component
