+++ client/routes.js
@@ -29,7 +29,7 @@ class Routes extends Component {
             <Route path="/login" component={Login} />
             <Route path="/signup" component={Signup} />
             
-            <Route path ="/users/:id" component={UserProfile} />
+            <Route path ="/users/:id" component={UserAccount} />
             {
               isLoggedIn &&
               <Switch>
@@ -37,6 +37,7 @@ class Routes extends Component {
                 <Route path="/activities" component={AllActivities} />
                 <Route path="/home" component={UserHome} />
                 <Route path="/uploadActivity" component={NewActivity} />
+                <Route path="/profile/:id" component={UserProfile} />
               </Switch>
             }
             {/* Displays our Login component as a fallback */}
