+++ client/store/index.js
@@ -5,7 +5,7 @@ import { composeWithDevTools } from 'redux-devtools-extension';
 import user from './user';
 import activity from './activity';
 import activities from './activities';
-import friends from './friends';
+import followers from './followers';
 import suggested from './suggested';
 
 const reducer = combineReducers({user, activity, activities, friends, suggested});
@@ -19,5 +19,5 @@ export default store;
 export * from './user';
 export * from './activity';
 export * from './activities';
-export * from './friends';
+export * from './followers';
 export * from './suggested';
